requirements.txt
mysql-replication==0.21
PyMySQL==0.9.3
PyYAML==5.3
redis==3.4.1
